fix soname of libgerbil on darwin
The soname/identification of libgerbil on darwin is not correctly fixed up,
resulting in any binaries linked with libgerbil to attempt to link to
libgerbil in the temporary build dir from the gerbil derivation. This
means that any compiled gerbil program won't work as the propagated soname
will result in the binary being unable to correctly link with libgerbil.

Adding an Darwin-only step to use `install_name_tool` to fix up the soname
of libgerbil fixes this issue. We're using realpath here as the qualified
path to `libgerbil` is `$out/gerbil/$version/lib/libgerbil.so` and this is
an easier way to obtain the qualified path.

Fixes NixOS/nixpkgs#272967.